Refactor and bug fixes

+++ src/main/scala/obb/evaluators/BoardEvaluator.scala
@@ -4,5 +4,5 @@ package obb.evaluators
 import obb.engine.{Board, Player}
 
 trait BoardEvaluator {
-  def evaluate(board : Board, player : Player) : Int
+  def evaluate(board : Board, player : Player) : Float
 }
